char to int

+++ 0x05-pointers_arrays_strings/holberton.h
@@ -21,6 +21,6 @@ void print_array(int *a, int n);
 
 char *_strcpy(char *dest, char *src);
 
-long int _atoi(char *s);
+int _atoi(char *s);
 
 #endif
